+++ spring-cloud-starter-tencent-polaris-router/src/test/java/com/tencent/cloud/polaris/router/endpoint/PolarisRouterEndpointAutoConfigurationTests.java
@@ -20,8 +20,9 @@ package com.tencent.cloud.polaris.router.endpoint;
 import com.tencent.cloud.polaris.context.ServiceRuleManager;
 import com.tencent.polaris.api.core.ConsumerAPI;
 import com.tencent.polaris.client.api.SDKContext;
+
 import com.tencent.polaris.factory.api.DiscoveryAPIFactory;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import org.springframework.boot.autoconfigure.AutoConfigurations;
 import org.springframework.boot.test.context.runner.WebApplicationContextRunner;
@@ -37,9 +38,12 @@ import static org.assertj.core.api.Assertions.assertThat;
  */
 public class PolarisRouterEndpointAutoConfigurationTests {
 
+	private final WebApplicationContextRunner contextRunner = new WebApplicationContextRunner()
+			.withConfiguration(AutoConfigurations.of(
+					TestServiceRuleManagerConfiguration.class,
+					PolarisRouterEndpointAutoConfiguration.class))
+			.withPropertyValues("endpoints.polaris-router.enabled=true");
 	private ServiceRuleManager serviceRuleManager;
-
-	private final WebApplicationContextRunner contextRunner = new WebApplicationContextRunner().withConfiguration(AutoConfigurations.of(TestServiceRuleManagerConfiguration.class, PolarisRouterEndpointAutoConfiguration.class)).withPropertyValues("endpoints.polaris-router.enabled=true");
 
 	@Test
 	public void polarisRouterEndpoint() {